+++ examples/js/example.js
@@ -1,5 +1,5 @@
 /* jshint browser: true */
-/* global require, angular, CONFIG */
+/* global require, angular, CONFIG, console */
 
 /**
  * @fileOverview
@@ -9,27 +9,29 @@
 
 'use strict';
 
-/** Module Dependencies */
-
-require('goangular');
-
 // Create an AngularJS application module
 var app = angular.module('app', ['goangular']);
+
 app.config(function($goConnectionProvider) {
   $goConnectionProvider.$set(CONFIG.connectUrl);
+  $goConnectionProvider.$loginUrl(null);
+  $goConnectionProvider.$loginUrl(['GitHub', 'Twitter', 'Salesforce']);
+  $goConnectionProvider.$logoutUrl();
 });
 
-app.controller('sweetController', function($scope, $goKey, $goConnection) {
+app.controller('sweetController', function($scope, $goQuery, $goUsers, $goConnection) {
+  window.scope = $scope; // expose scope for debugging
 
-  $scope.todos = $goKey('todos').$sync();
+  // $goConnection
+  $scope.conn = $goConnection;
 
-  $goConnection.$ready().then(function(connection) {
-    var room = connection.room('lobby');
-    room.self().get().then(function(results) {
-      $scope.currentUser = results.value;
-      $scope.$apply();
-    });
+  console.log($scope.conn.isGuest, 'isGuest'); // null
+  $scope.conn.$ready().then(function() {
+    console.log($scope.conn.isGuest, 'isGuest'); // boolean
   });
+
+  // $goKey
+  $scope.todos = $goQuery('todos', { sort: { '$name': 'desc'}}).$sync();
 
   $scope.addTodo = function() {
     var desc = $scope.newTodo;
@@ -51,14 +53,32 @@ app.controller('sweetController', function($scope, $goKey, $goConnection) {
     local: true
   };
 
-  $scope.todos.$on('add', opts, function(value) {
-    $scope.description = value.description;
+  $scope.todos.$on('add', opts, function(value, context) {
+    $scope.last = {
+      description: value.description,
+      user: $scope.users[context.userId].displayName
+    };
   });
 
   $scope.remove = function(key) {
     console.log('key remove is called on', key);
     $scope.todos.$key(key).$remove();
   };
+
+  // $goUsers
+  $scope.users = $goUsers().$sync();
+
+  $scope.users.$self(false).then(function(model) {
+    $scope.local = model.$sync();
+  });
+
+  $scope.users.$on('join', function(user) {
+    console.log('user joined', user.id, user.displayName);
+  });
+
+  $scope.users.$on('leave', function(user) {
+    console.log('user left', user.id, user.displayName);
+  });
 });
 
 app.directive('enter', function() {