+++ index.js
@@ -12,11 +12,16 @@
 
 var connectionFactory = require('./lib/connection_factory');
 
-var keySync = require('./lib/key_sync');
+var keySyncFactory = require('./lib/key_sync_factory');
 var keyFactory = require('./lib/key_factory');
 
 var querySync = require('./lib/query_sync');
 var queryFactory = require('./lib/query_factory');
+
+var usersSyncFactory = require('./lib/users_sync_factory');
+var usersFactory = require('./lib/users_factory');
+
+var keyFilter = require('./lib/key_filter');
 
 /** Module Registration */
 
@@ -26,7 +31,7 @@ var goangular = angular.module('goangular', []);
 
 goangular.provider('$goConnection', connectionFactory);
 
-goangular.factory('$goKeySync', [ '$parse', '$timeout', keySync ]);
+goangular.factory('$goKeySync', [ '$parse', '$timeout', keySyncFactory ]);
 goangular.factory('$goKey', [ '$goKeySync', '$goConnection', keyFactory ]);
 
 goangular.factory('$goQuerySync', [ '$parse', '$timeout', querySync ]);
@@ -36,3 +41,14 @@ goangular.factory('$goQuery', [
   '$goConnection',
   queryFactory
 ]);
+
+goangular.factory('$goUsersSync', [ '$parse', '$timeout', usersSyncFactory ]);
+goangular.factory('$goUsers', [
+  '$goUsersSync',
+  '$goKey',
+  '$goConnection',
+  '$q',
+  usersFactory
+]);
+
+goangular.filter('keyFilter', keyFilter);